20200423
1. tool 추가 (Scene)

2. tool에서 몬스터배치, 삭제, 저장, 불러오기 구현

3. 오브젝트 매니저 이름 없이 추가하는 메서드 오류 발견 후 수정

4. 저장한 데이터 메인씬에 불러오기 추가

+++ Project_Beom/Project_Beom/CameraManager.h
@@ -8,7 +8,7 @@ public:
 	virtual ~CameraManager();
 
 public:
-	void SetPos(int x, int y);
+	void SetPos(float x, float y);
 
 public:
 	const POSITION& GetPos() { return m_Pos; }
